fix(security): 整数演算精度損失警告解決 - GitHub Security Alert #1249
- Math.Sqrt計算前に整数をdoubleに明示的キャスト
- 距離計算での精度損失を防止
- オーバーフロー可能性を回避
- セキュリティスキャン警告 #1249 完全解決

+++ Baketa.Infrastructure/OCR/StickyRoi/InMemoryStickyRoiManager.cs
@@ -608,7 +608,7 @@ public sealed class InMemoryStickyRoiManager : IStickyRoiManager, IDisposable
         var dx = center1.X - center2.X;
         var dy = center1.Y - center2.Y;
         
-        return (int)Math.Sqrt(dx * dx + dy * dy);
+        return (int)Math.Sqrt((double)(dx * dx) + (double)(dy * dy));
     }
 
     private Rectangle AdjustRegionBounds(Rectangle current, Rectangle detected)
